backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
httpx
supabase-py
python-multipart
